% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cobalt_project.R
\name{create_cobalt_pro_project}
\alias{create_cobalt_pro_project}
\title{Creates a directory of folders for the project.}
\usage{
create_cobalt_pro_project(path = getwd())
}
\arguments{
\item{path}{Path of current working directory.}
}
\description{
Creates a directory of folders for the project.
}
