% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cobalt_project.R
\name{create_cobalt_project}
\alias{create_cobalt_project}
\title{This package will create a function called create_cobalt_project()}
\usage{
create_cobalt_project(path = getwd())
}
\description{
This package will create a function called create_cobalt_project()
}
