% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_education_entry.R
\name{create_education_entry}
\alias{create_education_entry}
\title{Generates education entry.}
\usage{
create_education_entry(
  degree_type,
  degree_name,
  institution,
  start_date,
  end_date
)
}
\arguments{
\item{degree_type}{Type of degree (e.g., BSc, MSc, PhD, MD, etc.)}

\item{degree_name}{Area of study in which degree was obtained}

\item{institution}{university/colllege where degree was obtained}

\item{start_date}{start date}

\item{end_date}{end_date}
}
\value{
A character vector containing LaTeX code
}
\description{
Generates education entry.
}
\examples{
create_education_entry(degree_type = 'PhD', degree_name = 'Industrial-Organizational Psychology',
                     institution = 'University of Guelph', start_date = 'Sep. 2018', end_date = 'May 2023')
}
