% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resume_entry.R
\name{create_resume_entry}
\alias{create_resume_entry}
\title{Generates resume entries.}
\usage{
create_resume_entry(
  title,
  subtitle = "",
  date,
  first_point = "",
  second_point = "",
  third_point = "",
  fourth_point = "",
  fifth_point = ""
)
}
\arguments{
\item{title}{Title of the resume entry.}

\item{subtitle}{Additional information related to the title.}

\item{date}{Date of achievement/completion or period of time.}

\item{first_point}{First point (if necessary)}

\item{second_point}{Second point (if necessary)}

\item{third_point}{Third point (if necessary)}

\item{fourth_point}{Fourth point (if necessary)}

\item{fifth_point}{Fourth point (if necessary)}
}
\value{
A character vector containin LaTeX code
}
\description{
Generates resume entries.
}
\examples{
create_resume_entry(title = "\\\\href{https://github.com/sebsciarra/smltheory}{smltheory}",
                    subtitle = "Python package", date = "Aug. 2023",
                    first_point = "Package contains nine modules and 30 functions",
                    second_point = "Functions within package simulate data sets and demonstrate propositions
                                     of supervised machine learning propositions (e.g., bias-variance tradeoff,
                                     excess risk decomposition)")
}
