% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_skills_entry.R
\name{create_skills_entry}
\alias{create_skills_entry}
\title{Generates skills entry.}
\usage{
create_skills_entry(coding_languages, ides_platforms, technical_skills)
}
\arguments{
\item{coding_languages}{Character ector containing coding languages you have used.}

\item{ides_platforms}{Character cector containing IDEs/platforms you have used (e.g., RStudio, Python, MySQL, etc)}

\item{technical_skills}{Character vector containing technicals skills you know (e.g., supervised machine learning).}
}
\value{
A character vector
}
\description{
Generates skills entry.
}
\examples{
create_skills_entry(coding_languages = c("Python", "R", "SQL", "\\\\LaTeX", "Javascript", "HTML", "CSS"),
                    ides_platforms = c("PyCharm", "RStudio", "AWS", "Git", "MySQL"),
                    technical_skills = c("Data visualization (ggplot2, plotnine)",
                                         "Data cleaning (tidyverse, pandas, numpy)",
                                         "Machine learning (e.g., regularized regression, decision
                                          trees, random forests, mixture models)",
                                         "Statistics (e.g., latent variable models, factor analysis,
                                           multilevel modelling)"))
}
