% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_white_paper_entry.R
\name{create_white_paper_entry}
\alias{create_white_paper_entry}
\title{Generates selected white papers entry.}
\usage{
create_white_paper_entry(links, titles, dates)
}
\arguments{
\item{links}{Links to published papers/blogs.}

\item{titles}{Titles of published papers/blogs}

\item{dates}{Publication dates of papers/blogs.}
}
\value{
A character vector
}
\description{
Generates selected white papers entry.
}
\examples{
links = c("https://sebastiansciarra.com/technical_content/understanding_ml/",
          "https://sebastiansciarra.com/technical_content/em/",
          "https://sebastiansciarra.com/technical_content/mle/")
titles = c("The Game of Supervised Machine Learning: Understanding the Setup, Players, and Rules",
           "The Expectation-Maximization Algorithm: A Method for Modelling
            Mixtures of Distributions",
           "Probability, Likelihood, and Maximum Likelihood Estimation")
dates = c("10 August 2023",
          "28 April 2023",
          "19 March 2023")
create_white_paper_entry(links = links, titles = titles, dates = dates)
}
