% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_resume.R
\name{knit_resume}
\alias{knit_resume}
\title{Knits resume and provides option to merge resume with cover letter into one PDF file.}
\usage{
knit_resume(..., merge_resume_cover_letter = F)
}
\description{
@param merge_resume_cover_letter If TRUE, resume and cover letter files will be merged into one PDF file.
}
\details{
Function needs to be used to overwrite knitr::knit() by using knit: cobaltResume::knit_resume in the YAML header.
 @examples
knit: cobaltResume2.0::knit_resume(merge_resume_cover_letter = T)
}
