% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_resume_cover_letter.R
\name{merge_resume_cover_letter}
\alias{merge_resume_cover_letter}
\title{Merges résumé and cover letter into one PDF file.}
\usage{
merge_resume_cover_letter(...)
}
\description{
Function can be used to overwrite knitr::knit() by using knit: cobaltResume::merge_resume_cover_letter in the YAML header.
}
