% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resume_entry.R
\name{create_resume_entry}
\alias{create_resume_entry}
\title{Generates résumé entries.}
\usage{
create_resume_entry(
  title,
  subtitle = "",
  date,
  first_point = "",
  second_point = "",
  third_point = "",
  fourth_point = ""
)
}
\arguments{
\item{title}{Title of the résumé entry.}

\item{subtitle}{Additional information related to the title.}

\item{date}{Date of achievement/completion or period of time.}

\item{first_point}{First point (if necessary)}

\item{second_point}{Second point (if necessary)}

\item{third_point}{Third point (if necessary)}

\item{fourth_point}{Fourth point (if necessary)}
}
\value{
A character vector
}
\description{
Generates résumé entries.
}
\examples{
create_resume_entry(title = "PhD, Industrial-Organizational Psychology",
subtitle = "University of Guelph",
date = "Sep. 2018--May 2023",
first_point = "Thesis title: \\\\textbf{Is Timing Everything? The Effects of Measurement Timing on the \\\\newline Performance of Nonlinear Longitudinal Models}",
second_point = "Conducted Monte Carlo simulation experiments using R on an Amazon Web Services (AWS) instance to evaluate the performance of structured latent growth curve models under several conditions",
third_point = "My thesis can be viewed  \\\\href{\\\\detokenize{https://atrium.lib.uoguelph.ca/xmlui/bitstream/handle/10214/27543/Sciarra_Sebastian_202304_PhD.pdf?sequence=1&isAllowed=y}}{\\\\textcolor{blue}{here}}")
}
